+++ src/main/java/racing/RacingMain.java
@@ -1,9 +1,10 @@
 package racing;
 
 
+import racing.controller.RacingApplication;
 import racing.exception.RacingBaseException;
 import racing.generator.RandomNumberGenerator;
-import racing.input.RacingInput;
+import racing.controller.input.RacingInput;
 
 public class RacingMain {
     public static void main(String[] args) {
