+++ src/main/java/racing/view/output/RacingOutput.java
@@ -1,4 +1,4 @@
-package racing.output;
+package racing.view.output;
 
 public class RacingOutput {
     public void print(String output) {
