+++ src/test/java/racing/view/output/RacingOutputTest.java
@@ -1,9 +1,10 @@
-package racing.output;
+package racing.view.output;
 
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racing.view.output.RacingOutput;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
